Add test cases and change api for creating taskrunsets
TaskRunSet model is changed and holds the code and is_local itself
instead of pointing to another Task. So we can think of Task model as
some template tasks which occur more often and we can give the template code
to the api itself.
In order to run tests a celery worker should be running.